Added unit testing for the migration step
